pyproject.toml
[project]
name = "finalcheck"
version = "0.1.0"
description = "PDF compliance verification using LLMs for controlled generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.31.0",
    "PyPDF2>=3.0.0",
    "pillow>=10.0.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "PyMuPDF>=1.22.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
finalcheck = "finalcheck.app:main"
